Converts a dyadic data into a network in various formats

get_adjacency takes in a dyadic dataset and outputs an adjacency matrix representation.

Usage

get_adjacency(
  dyad_data,
  actor1 = NULL,
  actor2 = NULL,
  symmetric = TRUE,
  mode = "unipartite",
  weight = NULL,
  sum_dyads = FALSE,
  diag_to_NA = TRUE,
  missing_to_zero = TRUE
)

Arguments

dyad_data

a dyadic dataframe (or a tibble)

actor1

character: name of the actor 1 in dyad_data

actor2

character: name of the actor 2 in dyad_data

symmetric

logical: whether ties are symmetric, default is TRUE

mode

character: whether the network is unipartite or bipartite, default is unipartite

weight

character: name of the weighted edges variable, default is NULL

sum_dyads

logical: whether to sum up the weight value when there exists repeating dyads within the dataset

diag_to_NA

logical: whether diagonals should be set to NA, default is TRUE

missing_to_zero

logical: whether missing values should be set to zero, default is TRUE

Value

an adjacency matrix of class netify

Author

Ha Eun Choi, Cassy Dorff, Colin Henry, Shahryar Minhas

Examples

# load example directed event data from ICEWS
# this data comes in the form of a dyadic
# dataframe where all dyad pairs are listed
data(icews)

# subset to a particular year
icews <- icews[icews$year=='2010', ]

# generate a cross sectional, directed, and weighted network
# where the weights are verbCoop
icews_verbCoop <- get_adjacency(
  dyad_data=icews, actor1='i', actor2='j',
  symmetric=FALSE, weight='verbCoop' )
# generate a cross sectional, directed and weighted network
# where the weights are matlConf
icews_matlConf <- get_adjacency(
  dyad_data=icews, actor1='i', actor2='j',
  symmetric=FALSE, weight='matlConf' )