Longitudinal variants of get_adjacency() which returns a list of adjacency matrices

This function takes in a dyadic dataset and outputs a list of adjacency matrices in which the actor composition in the network can vary over time.

Usage

get_adjacency_list(
  dyad_data,
  actor1 = NULL,
  actor2 = NULL,
  time = NULL,
  symmetric = TRUE,
  mode = "unipartite",
  weight = NULL,
  sum_dyads = FALSE,
  actor_time_uniform = FALSE,
  actor_pds = NULL,
  diag_to_NA = TRUE,
  missing_to_zero = TRUE
)

Arguments

dyad_data

a dyadic dataframe (or a tibble)

actor1

character: name of the actor 1 variable in dyad_data

actor2

character: name of the actor 2 variable in dyad_data

time

character: name of the time variable in dyad_data, the values of the time variable itself should be numeric

symmetric

logical: whether ties are symmetric, default is TRUE

mode

character: whether the network is unipartite or bipartite, default is unipartite

weight

character: name of the weighted edges variable, default is NULL

sum_dyads

logical: whether to sum up the weight value when there exists repeating dyads within time periods

actor_time_uniform

logical: whether to assume actors are the same across the full time series observed in the data TRUE means that actors are the same across the full time series observed in the data FALSE means that actors come in and out of the observed data and their "existence" should be determined by the data, meaning that their first year of existence will be determined by the time point of their first event and their last year of existence by the time point of their last event

actor_pds

a data.frame indicating start and end time point for every actor, this can be created manually (see example) or using get_actor_time_info.R, if provided then choice of actor_time_uniform is irrelevant.

diag_to_NA

logical: whether diagonals should be set to NA, default is TRUE

missing_to_zero

logical: whether missing values should be set to zero, default is TRUE

Value

a list of adjacency matrices of class netify

Author

Cassy Dorff, Ha Eun Choi, Shahryar Minhas

Examples

# load example directed event data from ICEWS
# this data comes in the form of a dyadic
# dataframe where all dyad pairs are listed
data(icews)

# generate a longitudinal, directed, and weighted network
# where the weights are verbConf
# note that in longitudinal networks we can have all the 
# same actors in every year and if that's the case
# we set actor_time_uniform to TRUE, in the next
# example we'll show how to adjust when actors
# change over time
icews_verbConf <- get_adjacency_list(
  dyad_data=icews, 
  actor1='i', actor2='j', time='year',
  actor_time_uniform=TRUE,
  symmetric=FALSE, weight='verbConf' )
icews_verbConf
#> ✔ Hello, you have created network data, yay!
#> • Unipartite
#> • Asymmetric
#> • Weights from `verbConf`
#> • Longitudinal: 13 Periods
#> • # Unique Actors: 152
#> Network Summary Statistics (averaged across time):
#>           dens miss mean recip trans
#> verbConf 0.111    0  1.5 0.831 0.374
#> • Nodal Features: None
#> • Dyad Features: None