Batch processing version for multiple time periods

Batch processing version for multiple time periods

Usage

get_matrix_batch(
  n_rows_vec,
  n_cols_vec,
  actors_rows_list,
  actors_cols_list,
  matRowIndices_list,
  matColIndices_list,
  value_list,
  symmetric,
  missing_to_zero = TRUE,
  diag_to_NA = TRUE
)

Arguments

n_rows_vec

integer vector of number of rows for each matrix

n_cols_vec

integer vector of number of columns for each matrix

actors_rows_list

list of character vectors for row actors

actors_cols_list

list of character vectors for column actors

matRowIndices_list

list of integer vectors for row indices

matColIndices_list

list of integer vectors for column indices

value_list

list of numeric vectors for values

symmetric

logical indicating whether ties are symmetric

missing_to_zero

logical indicating whether missing values should be set to zero

diag_to_NA

logical indicating whether diagonal should be set to NA

Value

a list of adjacency matrices

Author

Shahryar Minhas, Ha Eun Choi