Convert igraph, network, socio-matrices/arrays, and lists of each of those object to a netify object

prep_for_netify takes in a network object (e.g., an igraph or network object) and outputs a netify object. If the input is already a matrix/array/list of matrices, then it simply calls new_netify. Otherwise, it calls the appropriate internal helper (process_igraph or process_network) to extract adjacency matrix and any nodal/dyadic attributes, and then calls new_netify.

Usage

prep_for_netify(net_obj, weight = NULL, ...)

Arguments

net_obj

An R object (e.g., igraph, network, socio-matrix/array, or a list of these objects) that you want to convert to a netify object.

weight

Optional. Name of the weight attribute in net_obj to be used as the main edge weight in the netify object. Default is NULL. Important to add for igraph and network objects because they do not have a default weight.

...

Additional arguments passed to the internal processing functions.

Value

a netify object

Author

Cassy Dorff, Shahryar Minhas