Adjust Plotting Parameters and Data for Netify Visualization

This function prepares the plotting parameters and data from a netify object for visualization. It sets defaults for plotting parameters if not supplied by the user and modifies the node and edge data frames based on the provided parameters and the attributes of the netify object.

Usage

adjust_plot_args(plot_args, net_dfs, obj_attrs)

Arguments

plot_args

A list of user-defined plotting arguments that may include options for text overlap, edge curvature, edge angles, line segment parameters, and labeling specific actors.

net_dfs

A list containing nodal and edge data frames as returned by decompose_netlet. nodal_data should contain node information and edge_data should contain edge information.

obj_attrs

Attributes of the netify object, which should include details about whether the network is symmetric, its type (cross-sectional or longitudinal), and other relevant properties.

Value

A list containing the adjusted plot_args and modified net_dfs ready for plotting.

Details

The function checks and sets default values for various plot parameters such as:

check_overlap: Whether to check for overlap in text labels (default TRUE).

curve_edges: Whether to curve the edges in the plot (default FALSE).

edge_curvature: The curvature of the edges (default 0.5).

edge_angle: The angle of line segments (default 90 degrees).

edge_ncp: Number of control points for edge curvature (default 5).

edge_arrow: Default arrow type for directed edges. It also adjusts the edge and node data based on whether the network is directed or undirected, and handles specific labeling of actors and transparency settings based on edge weights.

Author

Cassy Dorff, Shahryar Minhas