Data preparation for plotting 'netify' objects

This function prepares the necessary data components for visualizing 'netify' objects. It processes network attributes and setups up layout and aesthetic parameters for subsequent plotting.

Usage

net_plot_data(netlet, plot_args = list())

Arguments

netlet

A 'netify' object, which contains the network data structured for analysis and visualization.

plot_args

A list of arguments that influence the layout and presentation of the network visualization, including: - point_layout: Optional, user-provided node layout; if not provided, layout will be generated based on layout parameter. - layout: Specifies the layout algorithm from 'igraph' to position the nodes if point_layout is not provided. Available options include "nicely", "fruchterman.reingold", "kamada.kawai", "random", "circle", "star", "grid", "graphopt", "sugiyama", "drl", "lgl", "bipartite", "tree", "randomly", "dh", "fr", "kk", "gem", and "mds". - remove_isolates: Logical; if TRUE, isolates will be removed from the plot. Default is TRUE. - static_actor_positions: Logical indicating whether to use static positions for actors. - add_edges: Logical; if TRUE, edges will be added to the plot. Default is TRUE. - curve_edges: Logical; if TRUE, edges will be curved. Default is FALSE. - add_points: Logical; if TRUE, points (nodes) will be plotted. Default is TRUE. - add_text: Logical; if TRUE, text annotations will be added. Default is FALSE. - add_label: Logical; if TRUE, labels will be added. Default is FALSE. - select_text: A vector of node names to specifically add text to; others will not have text. - select_label: A vector of node names to specifically add labels to; others will not have labels.

Value

A list containing plot_args, ggnet_params, and net_dfs which are used for setting up the plot: - plot_args: Adjusted plotting arguments including layout and graphical settings. - ggnet_params: Graphical parameters organized for nodes, edges, and labels. - net_dfs: Data frames for nodes and edges prepared for plotting.

Author

Cassy Dorff, Shahryar Minhas