Plotting function for actor level statistics for netify objects

plot_actor_stats takes in the output from summary_actor function and creates visualizations for actor level statistics. This function supports both cross-sectional and longitudinal data, providing visual insights into the distribution of statistics either across all actors or focusing on specific actors over time.

Usage

plot_actor_stats(
  summary_df,
  across_actor = TRUE,
  specific_actors = NULL,
  specific_stats = NULL
)

Arguments

summary_df

A dataframe produced by the summary_actor function containing actor-level or actor-time level statistics. The function expects the actor column to be labeled as "actor" and the time column to be labeled as "time" for longitudinal data.

across_actor

Logical; if TRUE, visualizations will focus on the distribution of statistics across actors. If FALSE, visualizations will focus on specific actors. Default is TRUE. If setting across_actor to FALSE it is highly recommended to specify ten or fewer specific actors.

specific_actors

Optional; a vector of specific actor names for which statistics will be plotted. When NULL, statistics for all actors are considered. This is only relevant if across_actor is set to FALSE.

specific_stats

Optional; a vector of specific statistics to plot. If NULL, all available statistics in the dataframe are used. Given the number of statistics that are calculated by default for some network types, choosing a few will make the plot more readable.

Value

A ggplot object representing the requested visualization, which can be further customized or printed.

Details

This function is capable of generating different types of plots based on the structure of the input data:

For cross-sectional data, it can generate violin plots to show the distribution of statistics across actors or bar plots to compare specific actors.

For longitudinal data, it can generate violin plots to show how statistics distribute over time across actors or line plots to track changes in statistics for specific actors over time.