Summary method to get graph level statistics for netify objects

summary.netify processes a netify object to calculate and return a data frame of graph-level statistics. This function is designed to work with both cross-sectional and longitudinal netify data structures, providing a comprehensive overview of network characteristics such as density, reciprocity, and standard deviations of sending and receiving effects.

Usage

# S3 method for netify
summary(object, ...)

Arguments

object

An object of class netify, which should have been created using the function get_adjacency. This object contains the network data structured for analysis.

...

Additional parameters which can include user-defined statistical functions. These functions should take a matrix as input and return a scalar value. They will be applied to each network slice individually if the netify object represents longitudinal data.

other_stats: A named list of functions that take a matrix and return additional actor-level statistics to be included in the output. Each function should accept a matrix as input and return a vector or single value per actor. This allows for the inclusion of custom metrics in the summary output.

Value

A data frame where each row represents the network-level statistics for a single network or a single time point in a longitudinal study. Depending on the network type and data attributes, the columns can include:

num_actors: Number of actors in the network - for bipartite networks number of row and column actors are reported separately.

density: The proportion of possible connections that are actual connections within the network.

num_edges: The total number of edges in the network (does not take edge weight into account).

mean_edge_weight: The average weight of edges in the network, provided only for weighted networks.

sd_edge_weight: The standard deviation of edge weights in the network, provided only for weighted networks.

prop_edges_missing: The proportion of potential edges that are missing.

min_edge_value and max_edge_value: The minimum and maximum edge weights observed in the network, provided only for weighted networks.

competition_row and competition_col (defaults to competition for undirected networks): Measures network competitiveness using the Herfindahl-Hirschman Index (HHI), defined as \(\sum_{i=1}^{n} (s_i)^2\), where \(s_i\) is the proportion of interactions by actor \(i\) and \(n\) is the total number of actors. The index ranges from 1/n (indicating high diversity and competitive interaction across actors) to 1 (one actor dominates all interactions). Refer to Dorff, Gallop, & Minhas (2023) for an application of this measure in conflict networks.

sd_of_row_means and sd_of_col_means: Standard deviations of the sending and receiving effects (row and column means). These statistics are meant to describe the variability in actor behavior across the network.

covar_of_row_col_means: The covariance between sending and receiving effects, always takes weights into account and is only calculated for unipartite networks.

reciprocity: The reciprocity of the network, defined as the correlation between the adjacency matrix and its transpose, always takes weights into account and is only calculated for unipartite networks.

transitivity: The overall transitivity or clustering coefficient of the network, reflecting the likelihood that two neighbors of a node are connected (calculated using transitivity function from igraph).

Details

This function simplifies the process of extracting key network statistics across potentially multiple networks within a netify object. It is capable of handling both weighted and unweighted networks and can adjust its calculations based on the nature of the network data (cross-sectional vs. longitudinal).

Author

Cassy Dorff, Shahryar Minhas

Examples

# load icews data
data(icews)

# create netlet
netlet = netify(
    dyad_data=icews, actor1='i', actor2='j',
    time = 'year', symmetric=FALSE, weight='verbCoop' )

# calculate default summary stats
summ_graph = summary(netlet)
head(summ_graph)
#>    net num_actors   density num_edges mean_edge_weight sd_edge_weight
#> 1 2002        152 0.3787034      8692         19.59123       143.8213
#> 2 2003        152 0.3871994      8887         19.36358       144.5982
#> 3 2004        152 0.4145173      9514         19.84986       137.7292
#> 4 2005        152 0.4071976      9346         19.79488       148.6667
#> 5 2006        152 0.4108139      9429         21.09136       160.0782
#> 6 2007        152 0.4243203      9739         21.89753       162.4166
#>   prop_edges_missing min_edge_value max_edge_value competition_row
#> 1         0.01315789              0           6003      0.04093116
#> 2         0.01315789              0           5937      0.04199822
#> 3         0.01315789              0           5141      0.03481011
#> 4         0.01315789              0           6561      0.04067046
#> 5         0.01315789              0           7579      0.04151400
#> 6         0.01315789              0           7125      0.03887627
#>   competition_col sd_of_row_means sd_of_col_means covar_of_row_col_means
#> 1      0.03813621        44.91530        43.04935              0.9946888
#> 2      0.03930188        45.07759        43.32783              0.9872959
#> 3      0.03262993        41.25503        39.63005              0.9923809
#> 4      0.03624472        45.20964        42.17312              0.9932157
#> 5      0.03697499        48.76301        45.48502              0.9909337
#> 6      0.03514325        48.67808        45.77854              0.9940837
#>   reciprocity transitivity
#> 1   0.9778217    0.6058952
#> 2   0.9632488    0.6072045
#> 3   0.9769563    0.6215978
#> 4   0.9804325    0.6215075
#> 5   0.9771928    0.6277829
#> 6   0.9783703    0.6330626

# add custom summary stat
spinglass_ig = function(mat){
    g = prep_for_igraph(mat)
    comm = igraph::cluster_spinglass(g)
    num_comm = length(comm$csize)
    return( c(
        num_comm=num_comm,
        comm_modul = comm$modularity
    ) )
}

# since calculating communities can be intensive
# lets take subset of time periods
sub_net = subset_netlet(netlet, when_to_subset = as.character(2013:2014))

# feed custom summary stat into summary
summary(sub_net, 
    other_stats=list(spinglass_ig=spinglass_ig)) 
#>    net num_actors   density num_edges mean_edge_weight sd_edge_weight
#> 1 2013        152 0.4318142      9911         17.21301       130.4925
#> 2 2014        152 0.4258452      9774         18.39291       136.5412
#>   prop_edges_missing min_edge_value max_edge_value competition_row
#> 1         0.01315789              0           6320      0.04058869
#> 2         0.01315789              0           6327      0.04058415
#>   competition_col sd_of_row_means sd_of_col_means covar_of_row_col_means
#> 1      0.03642385        39.26572        36.78299              0.9939394
#> 2      0.03521224        41.95447        38.49827              0.9904402
#>   reciprocity transitivity spinglass_ig.num_comm spinglass_ig.comm_modul
#> 1   0.9782226    0.6314429                     7             0.003979870
#> 2   0.9730745    0.6297063                     5             0.003912803