Summary method to get actor level statistics for netify objects

summary_actor provides detailed actor-level statistics for netify objects, handling different network structures and weight conditions. It produces a data frame summarizing various network metrics like degree, strength, closeness, betweenness, and centrality measures.

Usage

summary_actor(netlet, invert_weights_for_igraph = TRUE, other_stats = NULL)

Arguments

netlet

Object of class netify, typically produced by get_adjacency or other network creation functions within the package.

invert_weights_for_igraph

Logical; if TRUE, the weights of the edges are inverted before being used in the calculation of closeness or betweenness centrality. This is because igraph treats edge weights as distances. Inverting weights can be crucial when higher weights should imply stronger (or more valuable) connections rather than longer distances. Default is TRUE.

other_stats

A named list of functions that take a matrix and return additional actor-level statistics to be included in the output. Each function should accept a matrix as input and return a vector or single value per actor. This allows for the inclusion of custom metrics in the summary output.

Value

A data.frame object summarizing actor-level statistics of the network(s). Depending on the structure and attributes of the netify object, the output includes:

Symmetric Unweighted: Degree, proportion of ties, closeness, betweenness, and eigenvector centrality.

Symmetric Weighted: Degree, proportion of ties, sum and average of strengths, closeness (weighted and normalized), betweenness (weighted and normalized), and eigenvector centrality.

Asymmetric Unweighted: In-degree, out-degree, total degree, in-propensity of ties, out-propensity of ties, total propensity of ties, in-closeness, out-closeness, total closeness, in-betweenness, out-betweenness, total betweenness, authority scores, and hub scores.

Asymmetric Weighted: Similar metrics as asymmetric unweighted but includes strength sums and averages, and all closeness and betweenness measures are weighted. Metrics are calculated based on the network type (symmetric or asymmetric) and whether the network is weighted.

Details

The function automatically adjusts calculations based on network symmetry and weight attributes, offering tailored statistical outputs for comprehensive network analysis. It supports handling of both cross-sectional and longitudinal network data, ensuring that each actor's metrics are accurately computed over time if applicable. Examples of additional computations (like authority or hub scores) are provided only for asymmetric networks.

Author

Cassy Dorff, Shahryar Minhas

Examples

# load icews data
data(icews)

# create a netify object
netlet = netify(
    dyad_data=icews, actor1='i', actor2='j',
    time = 'year',
    symmetric=FALSE, weight='verbCoop',
    mode='unipartite', sum_dyads=FALSE,
    actor_time_uniform=TRUE, actor_pds=NULL,
    diag_to_NA=TRUE, missing_to_zero=TRUE
)

# get actor-level statistics
actor_stats <- summary_actor(netlet)
head(actor_stats)
#>         actor time degree_in degree_out degree_total prop_ties_in prop_ties_out
#> 1 Afghanistan 2002        81         92          173    0.5364238     0.6092715
#> 2     Albania 2002        49         46           95    0.3245033     0.3046358
#> 3     Algeria 2002        65         68          133    0.4304636     0.4503311
#> 4      Angola 2002        67         64          131    0.4437086     0.4238411
#> 5   Argentina 2002        48         48           96    0.3178808     0.3178808
#> 6     Armenia 2002        66         71          137    0.4370861     0.4701987
#>   prop_ties_total strength_sum_in strength_sum_out strength_sum_total
#> 1               1            8694            10160              18854
#> 2               1             943             1022               1965
#> 3               1            1123             1205               2328
#> 4               1            1111             1311               2422
#> 5               1            1261             1427               2688
#> 6               1            2642             2665               5307
#>   strength_avg_in strength_avg_out strength_avg_total closeness_in
#> 1       57.576159        67.284768          62.430464     71.13409
#> 2        6.245033         6.768212           6.506623     42.48817
#> 3        7.437086         7.980132           7.708609     44.78256
#> 4        7.357616         8.682119           8.019868     51.35489
#> 5        8.350993         9.450331           8.900662     63.91481
#> 6       17.496689        17.649007          17.572848     65.51012
#>   closeness_out closeness_all betweenness_in betweenness_out betweenness_all
#> 1      68.44904      78.06787   1.324503e-02    1.324503e-02    1.324503e-02
#> 2      43.64733      47.16201   0.000000e+00    0.000000e+00    0.000000e+00
#> 3      39.95772      47.41426   0.000000e+00    0.000000e+00    0.000000e+00
#> 4      46.60123      54.85391   8.830022e-05    8.830022e-05    8.830022e-05
#> 5      61.26657      69.47750   0.000000e+00    0.000000e+00    0.000000e+00
#> 6      63.04273      71.46157   0.000000e+00    0.000000e+00    0.000000e+00
#>   authority_score   hub_score
#> 1      0.26866096 0.183233794
#> 2      0.01252200 0.009730031
#> 3      0.01406673 0.010671854
#> 4      0.01905067 0.011275109
#> 5      0.03580770 0.024732267
#> 6      0.05049222 0.039622725

# add statistic that get 
# the max incoming and outgoing tie
max_out <- function(mat){ apply(mat, 1, max, na.rm=TRUE) }
max_in <- function(mat){ apply(mat, 2, max, na.rm=TRUE) }
actor_stats_custom <- summary_actor(netlet, other_stats = list(max_out = max_out, max_in = max_in))
head(actor_stats_custom)
#>         actor time degree_in degree_out degree_total prop_ties_in prop_ties_out
#> 1 Afghanistan 2002        81         92          173    0.5364238     0.6092715
#> 2     Albania 2002        49         46           95    0.3245033     0.3046358
#> 3     Algeria 2002        65         68          133    0.4304636     0.4503311
#> 4      Angola 2002        67         64          131    0.4437086     0.4238411
#> 5   Argentina 2002        48         48           96    0.3178808     0.3178808
#> 6     Armenia 2002        66         71          137    0.4370861     0.4701987
#>   prop_ties_total strength_sum_in strength_sum_out strength_sum_total
#> 1               1            8694            10160              18854
#> 2               1             943             1022               1965
#> 3               1            1123             1205               2328
#> 4               1            1111             1311               2422
#> 5               1            1261             1427               2688
#> 6               1            2642             2665               5307
#>   strength_avg_in strength_avg_out strength_avg_total closeness_in
#> 1       57.576159        67.284768          62.430464     71.13409
#> 2        6.245033         6.768212           6.506623     42.48817
#> 3        7.437086         7.980132           7.708609     44.78256
#> 4        7.357616         8.682119           8.019868     51.35489
#> 5        8.350993         9.450331           8.900662     63.91481
#> 6       17.496689        17.649007          17.572848     65.51012
#>   closeness_out closeness_all betweenness_in betweenness_out betweenness_all
#> 1      68.44904      78.06787   1.324503e-02    1.324503e-02    1.324503e-02
#> 2      43.64733      47.16201   0.000000e+00    0.000000e+00    0.000000e+00
#> 3      39.95772      47.41426   0.000000e+00    0.000000e+00    0.000000e+00
#> 4      46.60123      54.85391   8.830022e-05    8.830022e-05    8.830022e-05
#> 5      61.26657      69.47750   0.000000e+00    0.000000e+00    0.000000e+00
#> 6      63.04273      71.46157   0.000000e+00    0.000000e+00    0.000000e+00
#>   authority_score   hub_score max_out max_in
#> 1      0.26866096 0.183233794    1516   1847
#> 2      0.01252200 0.009730031     160    157
#> 3      0.01406673 0.010671854      91    114
#> 4      0.01905067 0.011275109     147    176
#> 5      0.03580770 0.024732267     424    469
#> 6      0.05049222 0.039622725     547    609